+++ src/main/java/util/helpers/EvalFunc.java
@@ -2,5 +2,5 @@ package util.helpers;
 
 @FunctionalInterface
 public interface EvalFunc<T> {
-    int eval(T obj);
+    long eval(T obj);
 }
